advanced search as a function
